// src/app/model/Pokemon.ts
export class Pokemon {
  id: string;
  type: string;
  name: string;
  weight: string;
  base_experience: string;
  height: string;
  ability: string;
  hp: string;
  attack: string;
  defense: string;


  constructor() {
    this.id = '';
    this.type = '';
    this.name = '';
    this.weight = '';
    this.base_experience = '';
    this.height = '';
    this.ability = '';
    this.hp = '';
    this.attack = '';
    this.defense = '';
  }
}